// newcode/libs/modules/estimate/shared/src/lib/common/model/est-line-item-response-entity.interface.ts
import { IFilterResult } from '@libs/platform/common';
import { IEstLineItem2CostGroupEntity, IEstLineItemEntity } from '@libs/estimate/interfaces';
import { IEstimateCompositeEntity } from './estimate-composite-entity.interface';
import { IProjectEntity } from '@libs/project/interfaces';
import { IEstStructureDetailBaseEntity } from './est-structure-detai-basel-entity.interface';
import { IBasicsCustomizeQuantityTypeEntity, IBasicsCustomizeSystemOptionEntity } from '@libs/basics/interfaces';
import { IEstRoundingConfigComplete } from './est-rounding-config-complete.interface';
import { IEstAllowanceBaseEntity } from './est-allowance-entity.base.interface';
import { IEstAllMarkup2costcodeBaseEntity } from './est-all-markup-2costcode-entity.base.interface';
import { IEstCopyOptionBaseEntity } from './est-copy-option-entity.base.interface';
import { BasicsCostGroupComplete } from '@libs/basics/shared';

export interface IEstLineItemResponseEntity{
	FilterResult: IFilterResult;
	dtos: IEstLineItemEntity[] | null;
	prjEstComposites: IEstimateCompositeEntity[]|null;
	selectedPrj: IProjectEntity|null;
	isEstDynamicColumnActive?: boolean;
	dynamicColumns?: object|null;
	EstStructureDetails?: IEstStructureDetailBaseEntity[]|null;
	companyMdcContextFk: number;
	LookupAssemblies?: IEstLineItemEntity[]|null;
	IsHeaderStatusReadOnly : boolean;
	advancedAllowanceCc : number|null;
	SystemOptions : IBasicsCustomizeSystemOptionEntity[]|null;
	jobFk?: number|null;
	ExtendColumns?: object[]|null;
	LineItem2CostGroups?: IEstLineItem2CostGroupEntity[]|null;
	CostGroupCats?: BasicsCostGroupComplete|null;
	IsWQReadOnly?: boolean;
	IsTotalAQBudget?: boolean;
	DoConsiderDisabledDirect?: boolean;
	EstLineItemStatusList: { [key: number]: boolean };
	packageReferenceLen?: number|null;
	IsCalcTotalWithWQ: boolean;
	EstRoundingConfigDetails?: IEstRoundingConfigComplete|null;
	allowanceEntity?: IEstAllowanceBaseEntity | null;
	estMarkup2CostCodes?: IEstAllMarkup2costcodeBaseEntity[]|null;
	mdcCostCodeStructure:{ Id: number, ParentFk: number | null }[];
	prjCostCodeStructure:{ Id: number, ParentFk: number | null, MdcCostCodeFk: number | null }[];
	IsBudgetEditable?: boolean;
	advancedAll:number;
	lineItemsWithAdvancedAll: IEstLineItemEntity[]|null;
	fixedPriceLineItems: IEstLineItemEntity[]|null;
	QuantityTypes?: IBasicsCustomizeQuantityTypeEntity[]|null;
	IsFixedBudgetTotal: boolean;
	IsAllowAssemblyTemplateNavigation?: boolean|null;
	EstCopyOptions?: IEstCopyOptionBaseEntity | null;
	Boq2CalcQtySplitMap?: {Item1:number, Item2:number, Item3:boolean}[] | null;
	EquipmentAssemblyCostUnitAlwaysEditable:boolean;
}